Move async scan_files_filtered() into appropriate async_file module
